backend/app/requirements.txt
alembic==1.14.0
async-timeout==5.0.1
blinker==1.9.0
boto3==1.35.68
botocore==1.35.68
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
s3transfer==0.10.4
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.1.3
zipp==3.21.0
